[READY]Обновление конроллера и не только (#1174)
* Initial SS update. Removing of dynamic waits and etc.
- Air SS look refactor
- Start and Stop processing unification
- stoplag() proc improvements
- FEA files deletion

* Space Drift SS

* Orbit SS

* Some SS improvements and other
- В Air SS CHECK_TICK заменил на `if (MC_TICK_CHECK) return` потому что в `fire()` проке паузы нужно делать именно таким образом
- Замены в некоторых местах Remove() в списках на -=
- Дальнейшая унификация START и STOP PROCESSING где это уместно
- Перенос повторяющихся дефайнов из Air SS и доков по засу в файл где им и место

* Asset SS and assets caching system
- Удалил кучу неиспользуемых файлов из папки HTML, а также файлы, которые посылались пользователю, но нигде не использовались (pda иконки)
- Небольшой рефактор бумаги. Теперь показывает содержимое через прок `show_content()`
- Немного поменялась структура папки nanoui

* some text fix

* missing span classes

* Throwing SS and Callbacks
- Убраны все найденные спавны перед проком `throw_at()`, т.к. теперь эта процедура является неблокирующей

* Addtimer with using of callbacks

* Addtimer update and MC improvements

* Removing of object pooling system
- https://github.com/tgstation/tgstation/pull/23150

* some tweaks
- Использование typecacheof списка в местах, где типов объектов много - очень неплохо улучает проивзодительность

* START/STOP_PROCESSING правки

* Фаерлоки теперь процессятся только в закрытом состоянии
- Список постоянно процесящейся машинерии стал меньше примерно на 1200 объектов.

* Небольшой твик на динамический глайд сайз. Перемещение будет не таким рваным

* Увеличена скорость обновления освещения

* Revert "Небольшой твик на динамический глайд сайз. Перемещение будет не таким рваным"

This reverts commit f46e1fa94291e72033dfbade0957600eb86f5d2e.

* Добавление макросов INVOKE_ASYNC и QDEL_IN
- Очень частичное их применение (просто чтобы положить этому начало)

*  Lighting things
- `update_overlay()` optimization from https://github.com/d3athrow/vgstation13/pull/11927
- Faster lighting init

* Фикс проблем с броском у билдмода и мусоровыбрасывателей

* небольшие правки

* Убрана передача 4-м аргументом автомата в проке `throw_at`

* runtime fix
```
runtime error: undefined variable /obj/structure/lattice/var/buckled
proc name: load (/obj/vehicle/space/spacebike/load)
  source file: spacebike.dm,55
```

* runtime fix
```
runtime error: undefined proc or verb /obj/item/weapon/gun/projectile/m79/playsound().
proc name: InvokeAsync (/datum/callback/proc/InvokeAsync)
  source file: callback.dm,79
```

* fix #1214

* фикс рантаймов связанных с двойным вызовом `throw_impact()` прока
Происходило из-за одного вызова throw_impact из под Bump() и другого из под подсистемы

* некоторые вызовы `INVOKE_ASYNC` заменены на `waitfor` настройку
пересмотрел их применение в некоторых местах и думаю, что в таких ситуативных проках настройка смотрится уместней

* фикс невосполняющей голод еды
при куделе все активные таймеры объекта удаляются, поэтому таймер на переваривание не давал резуальтат

* временный фикс утечек памяти из-за нового метода обновления освещения https://github.com/tgstation/tgstation/pull/24754
временный, потому что это бага бьёнда, которую люмокс уже пофиксил в версии 1377, поэтому ждёмс

* фикс теслы, очень быстро съедающей поле
честно говоря не совсем улавливаю связь между теми изменениями, что были сделаны в коде и этим, но да ладно.
- теперь, если всё сделать как сказана на вики, то поле будет стабильно удерживать теслу с 5 шарами. для того, чтобы получить лучший результат нужно или генератор контролировать, или эмиттеры добавлять, или ещё что-нибудь
- вернул возможность увидеть теслу в темноте

* не удержался и сделал небольшой рефактор. также фикс рантаймов, которые возникали при отключении поля.
в основном изменения визуальные, но есть пара правок, типа ленивых списков полей и гереторов в `field_generator`, убирания нескольких спавнов и по мелочи. функционал измениться не должен был

* Небольшие изменений связанные с qdel() и его логикой
- Добавлен gc_destroyed состояние: GC_CURRENTLY_BEING_QDELETED. Выставляется объекту перед выполнением `Destroy()`. То есть, можно сказать, почти сразу после начала процесса удаление объекта через `qdel()`. Таким образом можно проверкой на это состояние избежать повторых вызовов `qdel()` процедуры, если объект не успел выполнить свой дестрой до конца и не перешёл на следующую стадию удаления.
- Добавлен макрос с проверкой на GC_CURRENTLY_BEING_QDELETED
-` qdeleted()` переделан в макрос

* Правки к рефактору поля и генератора поля

+++ code/game/machinery/vending.dm
@@ -82,8 +82,9 @@
 	return
 
 /obj/machinery/vending/Destroy()
-	qdel(coin)
-	coin = null
+	if(coin)
+		qdel(coin)
+		coin = null
 	return ..()
 
 /obj/machinery/vending/ex_act(severity)
@@ -220,7 +221,7 @@
 					stat &= ~NOPOWER
 					set_light(light_range_on, light_power_on)
 
-	else if(currently_vending && istype(W, /obj/item/device/pda) && W.GetID()) //ѕда с картой
+	else if(currently_vending && istype(W, /obj/item/device/pda) && W.GetID())
 		var/obj/item/weapon/card/I = W.GetID()
 		scan_card(I)
 
@@ -657,9 +658,8 @@
 		break
 	if (!throw_item)
 		return 0
-	spawn(0)
-		throw_item.throw_at(target, 16, 3, src)
-	src.visible_message("\red <b>[src] launches [throw_item.name] at [target.name]!</b>")
+	throw_item.throw_at(target, 16, 3)
+	visible_message("<span class='danger'>[src] launches [throw_item.name] at [target.name]!</span>")
 	return 1
 
 /obj/machinery/vending/proc/isWireColorCut(wireColor)
